test_zivid_camera.cpp: Log the end of tests
It can be hard to debug failing tests because all of the logging
output from the zivid_camera driver for all of the tests is outputted.
Previously we printed the start of each test to make it easier to
follow. In this commit I also print the end of each test, as well
as print a line of ----- characters to make each test more visibly
separated in the output.