Update zivid_camera test to use new SDK 2.9 file camera
In SDK 2.10 FileCameraZividOne.zfc will not work. Update to use the new
file camera for Zivid 2 M70, which works for SDK 2.9 and later.

This means that the tests will no longer pass for SDK 2.8 and earlier,
and that officially we don't support SDK 2.8 or earlier with zivid-ros
master branch. I consider that acceptable, as the alternative is to
have to support two file cameras in the tests, and 2.9 has been out
for about 4 months now.