Switch to using modern OpenCL CPU runtime for testing
The CI system depended on a severely outdated (Ubuntu 16.04 era) Intel
CPU runtime for OpenCL from an archive file on Dropbox. In fact the
classic Intel CPU runtime for OpenCL has been end-of-life for a life, in
favor of the new OneAPI stack.

This commit updates the CI system to use the OpenCL runtime from the
OneAPI stack from Intel's official repositories. This should allow us to
add support for newer OSs since the old package was getting too old for
that. We also then no longer have to depend on some file from Dropbox.

This commit was taken from the zivid-python repo, commit
b1be7bce34cb609a.