test_zivid_camera.cpp: Do not test for exact count of NaN's
This checked that the number of NaN's in the point cloud is exactly X.
However, X can vary a tiny bit between compute devices/drivers etc.
Below in this test we compare the ROS output from the output from
capturing with Zivid SDK directly. This should be good enough for
testing the ROS driver.

We keep a test that the point cloud does consists of some valid points,
to ensure that the ROI selected does not set everything to NaN.